+++ forui/lib/src/theme/theme_data.dart
@@ -125,6 +125,9 @@ final class FThemeData with Diagnosticable, FTransformable {
   /// The tile group's style.
   final FTileGroupStyle tileGroupStyle;
 
+  /// The time field's style.
+  final FTimeFieldStyle timeFieldStyle;
+
   /// The tooltip style.
   final FTooltipStyle tooltipStyle;
 
@@ -167,6 +170,7 @@ final class FThemeData with Diagnosticable, FTransformable {
     FTextFieldStyle? textFieldStyle,
     FTooltipStyle? tooltipStyle,
     FTileGroupStyle? tileGroupStyle,
+    FTimeFieldStyle? timeFieldStyle,
   }) {
     typography = typography ?? FTypography.inherit(colorScheme: colorScheme);
     style = style ?? FStyle.inherit(colorScheme: colorScheme, typography: typography);
@@ -224,6 +228,8 @@ final class FThemeData with Diagnosticable, FTransformable {
           tooltipStyle ?? FTooltipStyle.inherit(colorScheme: colorScheme, typography: typography, style: style),
       tileGroupStyle:
           tileGroupStyle ?? FTileGroupStyle.inherit(colorScheme: colorScheme, typography: typography, style: style),
+      timeFieldStyle:
+          timeFieldStyle ?? FTimeFieldStyle.inherit(colorScheme: colorScheme, typography: typography, style: style),
     );
   }
 
@@ -265,6 +271,7 @@ final class FThemeData with Diagnosticable, FTransformable {
     required this.textFieldStyle,
     required this.tooltipStyle,
     required this.tileGroupStyle,
+    required this.timeFieldStyle,
   });
 
   /// Converts this [FThemeData] to a Material [ThemeData] on a best-effort basis.
@@ -827,6 +834,7 @@ final class FThemeData with Diagnosticable, FTransformable {
     FTabsStyle? tabsStyle,
     FTextFieldStyle? textFieldStyle,
     FTileGroupStyle? tileGroupStyle,
+    FTimeFieldStyle? timeFieldStyle,
     FTooltipStyle? tooltipStyle,
   }) => FThemeData(
     colorScheme: colorScheme,
@@ -863,6 +871,7 @@ final class FThemeData with Diagnosticable, FTransformable {
     tabsStyle: tabsStyle ?? this.tabsStyle,
     textFieldStyle: textFieldStyle ?? this.textFieldStyle,
     tileGroupStyle: tileGroupStyle ?? this.tileGroupStyle,
+    timeFieldStyle: timeFieldStyle ?? this.timeFieldStyle,
     tooltipStyle: tooltipStyle ?? this.tooltipStyle,
   );
 
@@ -906,6 +915,7 @@ final class FThemeData with Diagnosticable, FTransformable {
       ..add(DiagnosticsProperty('tabsStyle', tabsStyle, level: DiagnosticLevel.debug))
       ..add(DiagnosticsProperty('textFieldStyle', textFieldStyle, level: DiagnosticLevel.debug))
       ..add(DiagnosticsProperty('tileGroupStyle', tileGroupStyle, level: DiagnosticLevel.debug))
+      ..add(DiagnosticsProperty('timeFieldStyle', timeFieldStyle, level: DiagnosticLevel.debug))
       ..add(DiagnosticsProperty('tooltipStyle', tooltipStyle, level: DiagnosticLevel.debug));
   }
 
@@ -949,6 +959,7 @@ final class FThemeData with Diagnosticable, FTransformable {
           tabsStyle == other.tabsStyle &&
           textFieldStyle == other.textFieldStyle &&
           tileGroupStyle == other.tileGroupStyle &&
+          timeFieldStyle == other.timeFieldStyle &&
           tooltipStyle == other.tooltipStyle;
 
   @override
@@ -988,5 +999,6 @@ final class FThemeData with Diagnosticable, FTransformable {
       tabsStyle.hashCode ^
       textFieldStyle.hashCode ^
       tileGroupStyle.hashCode ^
+      timeFieldStyle.hashCode ^
       tooltipStyle.hashCode;
 }