+++ forui/lib/widgets/time_field.dart
@@ -6,3 +6,4 @@
 library forui.widgets.time_field;
 
 export '../src/widgets/time_field/time_field.dart';
+export '../src/widgets/time_field/time_field_style.dart';